move "smashlike" to commons
